Evans : Code changes wrt Chad's comments.
